Web: Merge logic to get client IP address from CurrentUser and AuthenticationHander
Similar logic to get the client's IP address existed in both CurrentUser and
AuthenticationHander this commit moves the login into a RemoteAddressUtil class
