+++ web/web-base/src/main/java/org/visallo/web/AuthenticationHandler.java
@@ -8,7 +8,6 @@ import javax.servlet.http.HttpServletResponse;
 
 public class AuthenticationHandler implements RequestResponseHandler {
     public static final String LOGIN_PATH = "/login";
-    private static final String HEADER_X_FORWARDED_FOR = "x-forwarded-for";
 
     @Override
     public void handle(HttpServletRequest request, HttpServletResponse response, HandlerChain chain) throws Exception {
@@ -18,13 +17,5 @@ public class AuthenticationHandler implements RequestResponseHandler {
         } else {
             response.sendError(HttpServletResponse.SC_FORBIDDEN);
         }
-    }
-
-    public static String getRemoteAddr(HttpServletRequest request) {
-        String xForwardedFor = request.getHeader(HEADER_X_FORWARDED_FOR);
-        if (xForwardedFor != null && xForwardedFor.trim().length() > 0) {
-            return xForwardedFor;
-        }
-        return request.getRemoteAddr();
     }
 }
