+++ web/web-base/src/main/java/org/visallo/web/util/RemoteAddressUtil.java
@@ -1,0 +1,27 @@
+package org.visallo.web.util;
+
+import javax.servlet.http.HttpServletRequest;
+
+public class RemoteAddressUtil {
+    // code adapted from: http://stackoverflow.com/a/15323776/39431
+    public static String getClientIpAddr(HttpServletRequest request) {
+        String ip;
+        ip = request.getHeader("X-Forwarded-For");
+        if (ip == null || ip.length() == 0 || "unknown".equalsIgnoreCase(ip)) {
+            ip = request.getHeader("Proxy-Client-IP");
+        }
+        if (ip == null || ip.length() == 0 || "unknown".equalsIgnoreCase(ip)) {
+            ip = request.getHeader("WL-Proxy-Client-IP");
+        }
+        if (ip == null || ip.length() == 0 || "unknown".equalsIgnoreCase(ip)) {
+            ip = request.getHeader("HTTP_CLIENT_IP");
+        }
+        if (ip == null || ip.length() == 0 || "unknown".equalsIgnoreCase(ip)) {
+            ip = request.getHeader("HTTP_X_FORWARDED_FOR");
+        }
+        if (ip == null || ip.length() == 0 || "unknown".equalsIgnoreCase(ip)) {
+            ip = request.getRemoteAddr();
+        }
+        return ip;
+    }
+}
